+++ templates/srtf.html
@@ -3,7 +3,7 @@
 <head>
   <meta charset="UTF-8">
   <title>FCFS Scheduling Simulator</title>
-  <link rel="stylesheet" href="{{ url_for('static', filename='css/sjf_p.css') }}">
+  <link rel="stylesheet" href="{{ url_for('static', filename='css/srtf.css') }}">
 </head>
 <body>
   <div class="container">
@@ -34,6 +34,6 @@
     </div>
   </div>
 
-  <script src="{{ url_for('static', filename='js/sjf_p.js') }}"></script>
+  <script src="{{ url_for('static', filename='js/srtf.js') }}"></script>
 </body>
 </html>
